pyproject.toml
[project]
name = "paper-reading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "tenacity>=9.1.2",
    "types-pyyaml>=6.0.12.20250915",
]
